feat: rework of the package & Laravel 6 / 7 support
Thanks to @jbeales https://github.com/designmynight/laravel-log-mailer/pull/7
