+++ CHANGELOG.md
@@ -1,12 +1,21 @@
 # Changelog
 
-All notable changes to `laravel-log-mailer` will be documented in this file
+All notable changes to `laravel-mail-log-channel` will be documented in this file
+
+## 1.1.0 - 2020-01-29
+
+- add support for Laravel 6 and 7 (thanks to @jbeales https://github.com/designmynight/laravel-log-mailer/pull/7)
+- remove extra configuration and view files
+- improve exception layout in mails
 
 ## 1.0.2 - 2018-09-09
+
 - fix logging levels
 
 ## 1.0.1 - 2018-09-05
+
 - fix dependancy
 
 ## 1.0.0 - 2018-09-04
+
 - initial release
