+++ composer.json
@@ -1,24 +1,36 @@
 {
-    "name": "designmynight/laravel-log-mailer",
+    "name": "shaffe/laravel-mail-log-channel",
     "description": "A package to support logging via email in Laravel",
-    "homepage": "https://github.com/designmynight/laravel-log-mailer",
+    "homepage": "https://github.com/shaffe-fr/laravel-mail-log-channel",
     "license": "MIT",
     "keywords": [
         "laravel",
-        "laravel-logging",
+        "logging",
+        "mail channel",
         "laravel-log-mailer",
         "monolog",
-        "designmynight"
+        "laravel-mail-log-channel",
+        "shaffe"
     ],
     "require": {
-        "illuminate/support": "^5.6"
+        "illuminate/bus": "^5.6|^6.0|^7.0",
+        "illuminate/contracts": "^5.6|^6.0|^7.0",
+        "illuminate/log": "^5.6|^6.0|^7.0",
+        "illuminate/mail": "^5.6|^6.0|^7.0",
+        "illuminate/queue": "^5.6|^6.0|^7.0",
+        "illuminate/support": "^5.6|^6.0|^7.0"
     },
     "autoload": {
         "psr-4": {
-            "DesignMyNight\\Laravel\\": "src"
+            "Shaffe\\MailLogChannel\\": "src"
         }
     },
     "authors": [
+        {
+            "name": "Karel FAILLE",
+            "email": "shaffe.fr@gmail.com",
+            "role": "Developer"
+        },
         {
             "name": "Steve Porter",
             "email": "steve@designmynight.com",
@@ -28,7 +40,7 @@
     "extra": {
         "laravel": {
             "providers": [
-                "DesignMyNight\\Laravel\\Logging\\MailableLogServiceProvider"
+                "Shaffe\\MailLogChannel\\MailLogChannelServiceProvider"
             ]
         }
     }
