--- resources/views/log.blade.php
@@ -1,5 +1,0 @@
-@component('mail::message')
-
-{!!$content!!}
-
-@endcomponent
